+++ CHANGELOG.md
@@ -39,6 +39,9 @@
 
   * `dart:io`
     * Adds `HttpClient.connectionTimeout`.
+    * Adds `{Socket,RawSocket,SecureSocket}.startConnect`. These return a
+      `ConnectionTask`, which can be used to cancel an in-flight connection
+      attempt.
 
 ## 2.0.0-dev.66.0